+++ Assets/Script/Smin/Bullet/Bullet_Base.cs
@@ -7,7 +7,7 @@ public abstract class Bullet_Base : MonoBehaviour
     float cur_lifeTime;
     [SerializeField] protected float lifeTime;
     [SerializeField] protected float moveSpeed;
-    [SerializeField] protected int damage;
+    [SerializeField] protected float damage;
 
     [SerializeField] protected bool isStop;
 
@@ -15,7 +15,7 @@ public abstract class Bullet_Base : MonoBehaviour
         isStop = _isStop;
     }
 
-    public void Init(float _lifeTime, float _moveSpeed, int _damage)
+    public void Init(float _lifeTime, float _moveSpeed, float _damage)
     {
         lifeTime = _lifeTime;
         moveSpeed = _moveSpeed;
